Minor cleanups found while fixing a bug involving double mapping of low memory:
Load the kernel eflags less magically, as in locore.  The magic increased
when I removed eflags from the pcb in r305899.

Remove a jump to low memory that became garbage when the i386 version was
mostly replaced by the amd64 version in r235622.

The amd64 version is very similar.  It still loads the flags magically,
but is not missing comments about using the special page table.

Reviewed by:	kib


git-svn-id: https://svn.freebsd.org/base/head@326867 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f