Also forgotten in the previous that removed the permanent double mapping of low physical memory:
Update the comment about leaving the permanent mapping in place.  This
also improves the wording of the comment.  PTD 0 is still left alone
because it is fairly important that it was unmapped earlier, and the
comment now describes the unmapping of the other low PTDs that the code
actually does.

Reviewed by:	kib


git-svn-id: https://svn.freebsd.org/base/head@326933 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f