Make truss work for CloudABI executables on i386.
The system call convention is different from i386 binaries running on
FreeBSD/amd64, but this is not noticeable by executables. On
FreeBSD/amd64, the vDSO already does padding of arguments and return
values to 64-bit values. On i386, it does not, meaning that system call
return values are simply stored in registers.


git-svn-id: https://svn.freebsd.org/base/head@327057 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f