MFC r325108:
Improve the performance of the hpet timer in bhyve guests by making the
timer frequency a power of two.  This changes the frequency from 10 to
16.7 MHz (2 ^ 24 HZ).  Using a power of two avoids roundoff errors when
doing arithmetic in sbintime_t units.

Testing shows this can fix erratic ntpd behavior in guests using the
hpet timer (which is the default for multicore guests).


git-svn-id: https://svn.freebsd.org/base/stable/11@327651 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f