MFC r327220-r327221
r327220:
Update the FDT example for the lm75 sensor to match current devicetree
standards and what the existing driver expects.

Also change 'like' to 'such as' where the text is providing an example
rather than a simile.

r327221:
Complete the changing of the old "i2c-address" property to the modern "reg"
property by updating the description text to match the updated example.

This should have been part of r327220


git-svn-id: https://svn.freebsd.org/base/stable/11@327668 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f